requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.4
django-autoslug==1.9.9
django-crispy-forms==1.14.0
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.5.0
djangorestframework==3.14.0
pillow==10.4.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2022.7
